Move all resolver test helpers into same file
